unit1302 backtrack unrelated change

+++ tests/unit/unit1302.c
@@ -38,7 +38,7 @@ static CURLcode test_unit1302(const char *arg)
 {
   UNITTEST_BEGIN_SIMPLE
   CURLcode rc;
-  size_t i;
+  unsigned int i;
 
   /* common base64 encoding */
   struct etest encode[] = {
@@ -137,7 +137,7 @@ static CURLcode test_unit1302(const char *arg)
     abort_unless(rc == CURLE_OK, "return code should be CURLE_OK");
     abort_unless(olen == e->olen, "wrong output size");
     if(memcmp(out, e->output, e->olen)) {
-      curl_mfprintf(stderr, "Test %zu encoded badly\n", i);
+      curl_mfprintf(stderr, "Test %u encoded badly\n", i);
       unitfail++;
     }
     Curl_safefree(out);
@@ -145,16 +145,16 @@ static CURLcode test_unit1302(const char *arg)
     /* then verify decode */
     rc = curlx_base64_decode(e->output, &decoded, &dlen);
     if(rc != CURLE_OK) {
-      curl_mfprintf(stderr, "Test %zu URL decode returned %d\n", i, (int)rc);
+      curl_mfprintf(stderr, "Test %u URL decode returned %d\n", i, (int)rc);
       unitfail++;
     }
     if(dlen != e->ilen) {
-      curl_mfprintf(stderr, "Test %zu URL decode output length %zu "
+      curl_mfprintf(stderr, "Test %u URL decode output length %zu "
                     "instead of %zu\n", i, dlen, e->ilen);
       unitfail++;
     }
     if(memcmp(decoded, e->input, dlen)) {
-      curl_mfprintf(stderr, "Test %zu URL decoded badly. Got '%s', "
+      curl_mfprintf(stderr, "Test %u URL decoded badly. Got '%s', "
                     "expected '%s'\n", i, decoded, e->input);
       unitfail++;
     }
@@ -169,11 +169,11 @@ static CURLcode test_unit1302(const char *arg)
     rc = curlx_base64url_encode(e->input, e->ilen, &out, &olen);
     abort_unless(rc == CURLE_OK, "return code should be CURLE_OK");
     if(olen != e->olen) {
-      curl_mfprintf(stderr, "Test %zu URL encoded output length %zu "
+      curl_mfprintf(stderr, "Test %u URL encoded output length %zu "
                     "instead of %zu\n", i, olen, e->olen);
     }
     if(out && memcmp(out, e->output, e->olen)) {
-      curl_mfprintf(stderr, "Test %zu URL encoded badly. Got '%s', "
+      curl_mfprintf(stderr, "Test %u URL encoded badly. Got '%s', "
                     "expected '%s'\n", i, out, e->output);
       unitfail++;
     }
@@ -188,7 +188,7 @@ static CURLcode test_unit1302(const char *arg)
     /* then verify decode with illegal inputs */
     rc = curlx_base64_decode(e->output, &decoded, &dlen);
     if(rc != CURLE_BAD_CONTENT_ENCODING) {
-      curl_mfprintf(stderr, "Test %zu URL bad decoded badly. "
+      curl_mfprintf(stderr, "Test %u URL bad decoded badly. "
                     "Returned '%d', expected '%d'\n",
                     i, (int)rc, CURLE_BAD_CONTENT_ENCODING);
       unitfail++;
